+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [Issue 417](https://github.com/aza547/wow-recorder/issues/417) - Fix a bug where sometimes recordings were missed.
 - [NO-ISSUE] - Split out the chat overlay settings from the game settings so it's more responsive.
+- [Issue 418](https://github.com/aza547/wow-recorder/issues/418) - Improve the test button UX.
 
 ## [3.13.0] - 2023-05-21
 ### Added